Update README.md
The "Process Map Data using C++ Adjacency List Shortest Path" project, created by "Kushagra", is a C++ program that demonstrates the concept of finding the shortest paths in a graph. It employs an adjacency list representation, along with an "Edge" class to model graph edges, each with a target vertex and a weight. The program enables users to calculate and display the shortest distances from a source vertex to other vertices. It also provides a sample CSV file, "towns.csv," for graph data. This educational project serves as a practical introduction to graph algorithms and graph representation, offering a hands-on experience for algorithm development and understanding.